# apps/cockpit/openapi.json
{
  "components": {
    "schemas": {
...
      "ContractOut": {
        "properties": {
          "compliance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance"
          },
          "dataset": {
            "default": "",
            "title": "Dataset",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "guarantees": {
            "additionalProperties": true,
            "default": {},
            "title": "Guarantees",
            "type": "object"
          },
          "lifecycle": {
            "default": "draft",
            "title": "Lifecycle",
            "type": "string"
          },
          "owned_by": {
            "default": "platform",
            "title": "Owned By",
            "type": "string"
          },
          "owners": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Owners",
            "type": "array"
          },
          "product": {
            "title": "Product",
            "type": "string"
          },
          "version": {
            "default": "0.1.0",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "product"
        ],
        "title": "ContractOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "DQ & Observability Cockpit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/contracts": {
      "get": {
        "operationId": "list_contracts_api_contracts_get",
        "parameters": [
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContractOut"
                  },
                  "title": "Response List Contracts Api Contracts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Contracts",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{product}": {
      "get": {
        "operationId": "get_contract_api_contracts__product__get",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Contract",
        "tags": [
          "contracts"
        ]
      },
      "put": {
        "description": "[CONTRACT-SQL-FREE] [AUTHZ] Update a contract \u2014 Gate G1 enforced here.\n\nLifecycle ist kein Eingabefeld: das Ergebnis eines PUT ist IMMER ein Draft.\nEin PUT auf einen aktiven Contract erzeugt damit ein Draft-Amendment \u2014 die\nzertifizierte Version bleibt als `.active.yml`-Snapshot die G3-Diff-Basis,\nbis ein erneutes Approve sie ersetzt.",
        "operationId": "update_contract_api_contracts__product__put",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Contract",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{product}/approve": {
      "post": {
        "description": "[AUTHZ] Promote a draft contract to active (WS2-6 / M2).\n\nServer-side breaking guard (G3): if the change is breaking versus the last\ncertified version, a SemVer **major** bump is required.",
        "operationId": "approve_contract_api_contracts__product__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Contract",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{product}/compile": {
      "post": {
        "description": "[DETERMINISM] [SCHEMA-MAP] Compile contract \u2192 CheckDefs.\n\nG1: der Contract wird auch auf dem Disk-Pfad validiert (kein Schmuggel an\nPUT vorbei). G2: '{schema}' bleibt als Platzhalter im Output. Non-dry-run\nerfordert lifecycle=active (WS3-2) und Schreibrecht.",
        "operationId": "compile_contract_api_contracts__product__compile_post",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": false,
              "title": "Dry Run",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compile Contract",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{product}/deprecate": {
      "post": {
        "description": "[AUTHZ] Retire an active contract (WS2-6).",
        "operationId": "deprecate_contract_api_contracts__product__deprecate_post",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deprecate Contract",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{product}/diff": {
      "post": {
        "description": "Compare submitted contract against current version (WS2-4).",
        "operationId": "diff_contract_api_contracts__product__diff_post",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff Contract",
        "tags": [
          "contracts"
        ]
      }
    },
...
  }
}